rename: move from datasets._io to settings._io
